requirements.txt
pillow==9.3.0
pyautogui==0.9.54
keyboard==0.13.5
